move DisabledXvfbController to testing module
