+++ xvfb-testing/src/main/java/com/github/mike10004/xvfbtesting/XvfbRule.java
@@ -3,9 +3,7 @@
  */
 package com.github.mike10004.xvfbtesting;
 
-import com.github.mike10004.xvfbmanager.DisabledXvfbController;
 import com.github.mike10004.xvfbmanager.XvfbController;
-import com.github.mike10004.xvfbmanager.XvfbException;
 import com.github.mike10004.xvfbmanager.XvfbManager;
 import com.novetta.ibg.common.sys.Platforms;
 import org.junit.rules.ExternalResource;
